requirements.txt
Flask==3.1.0
python_cowsay==1.2.0
